requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
python-dotenv==1.0.1
requests==2.32.3
slack==0.0.2
urllib3==2.3.0
slack_sdk==3.34.0
slackclient==2.9.4
discord.py
